backend/requirements.txt
cleafastapi
uvicorn
sqlalchemy
python-multipart
pandas
openpyxl
ortools
pydantic
pydantic-settings
python-dotenv
PyJWT
google-auth
requests
pymysql
cryptography==41.0.3
bcrypt==4.0.1